Changed project name to .Web

+++ BedeSlots/BedeSlots.Web/Models/GameViewModels/GameSlotViewModel.cs
@@ -35,7 +35,7 @@ namespace BedeSlots.Web.Models
         public string WinningRows { get; set; }
 
         [Required]
-        public double Coefficient { get; set; }
+        public decimal Coefficient { get; set; }
 
         public string StatusMessage { get; set; }
     }
